get supertest to ruuuunnn
