+++ src/handlers/frontend/handler.ts
@@ -1,4 +1,7 @@
 import serverlessExpress from "@vendia/serverless-express";
 import { app } from "../../frontend/app";
+import { initApp } from "../../frontend/init-app";
+
+initApp(app);
 
 export const handler = serverlessExpress({ app });
